refactor(chat): use directly editor object in order to fix sync issues
